Added edit and delete widget functionality

+++ application/views/pages/edit/index.php
@@ -11,8 +11,6 @@
 	<div class="col-lg-7 col-lg-offset-1">
 		<? if(count($widgets) > 0) { ?>
 
-			<? $total_amount = 0; ?>
-
 			<table id="widget-table" class="table table-condensed">
 
 				<tr>
@@ -24,15 +22,13 @@
 				
 				<? foreach($widgets as $widget) { ?>
 
-					<? $total_amount+= $widget->value; ?>
-
 					<tr>
 						<td><?= $widget->name; ?></td>
 						<td><?= $widget->description; ?></td>
 						<td class="align-right">$<?= number_format($widget->value,2); ?></td>
 						<td class="align-right">
 							<a href="<?= base_url('edit/edit_item/'.$widget->id); ?>" class="btn btn-primary">Edit</a>
-							<a href="<?= base_url('edit/delete_item/'.$widget->id); ?>" class="btn btn-danger">Delete</a>
+							<a href="<?= base_url('edit/delete_item/'.$widget->id); ?>" class="btn btn-danger" onclick="return confirm('Are you sure you want to delete this item?');">Delete</a>
 						</td>
 					</tr>
 
@@ -41,12 +37,14 @@
 				<tr>
 					<td></td>
 					<td class="align-right">Total Cost:</td>
-					<td class="align-right">$<?= number_format($total_amount,2); ?></td>
+					<td class="align-right">$<?= number_format($totalValue,2); ?></td>
 					<td></td>
 				</tr>
 
 			</table>
 
+		<? } else { ?>
+			<span>No widgets in the system. Click <a href="<?= site_url('page-that-does-not-exist'); ?>">here</a> to add widgets</span>
 		<? } ?>
 	</div>
 </div>
